packages/agent-gateway/requirements.txt
# Agent Gateway — Python FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0
sse-starlette>=1.8.0
openai>=1.6.0
aiofiles>=23.2.0
